hailuo-clone/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
gradio-client==0.7.1
python-dotenv==1.0.0
requests==2.31.0
Pillow>=10.0.0
